Restructure iteration, retry, sessionparameters, transport, txparameters, versions
