+++ tests/stub/retry/scripts/commit_disconnect.script
@@ -1,0 +1,12 @@
+!: BOLT 4
+!: AUTO HELLO
+!: AUTO RESET
+
+C: BEGIN {}
+S: SUCCESS {}
+C: RUN "RETURN 1" {} {}
+   PULL {"n": 1000}
+S: SUCCESS {"fields": ["n"]}
+   SUCCESS {"type": "w"}
+C: COMMIT
+S: <EXIT>
