+++ tests/stub/retry/scripts/retry_with_fail_after_commit.script
@@ -1,0 +1,26 @@
+!: BOLT 4
+!: AUTO HELLO
+!: AUTO GOODBYE
+
+C: BEGIN {}
+S: SUCCESS {}
+C: RUN "RETURN 1" {} {}
+   PULL {"n": 1000}
+S: SUCCESS {"fields": ["n"]}
+   RECORD [1]
+   SUCCESS {"type": "r"}
+C: COMMIT
+S: FAILURE {"code": "#ERROR#", "message": "<whatever>"}
+C: RESET
+S: SUCCESS {}
+#EXTRA_RESET_1#
+C: BEGIN {}
+S: SUCCESS {}
+C: RUN "RETURN 1" {} {}
+   PULL {"n": 1000}
+S: SUCCESS {"fields": ["n"]}
+   RECORD [1]
+   SUCCESS {"type": "r"}
+C: COMMIT
+S: SUCCESS {}
+#EXTRA_RESET_2#
