+++ tests/stub/retry/scripts/retry_with_fail_after_pull_server1.script
@@ -1,0 +1,13 @@
+!: BOLT 4
+!: AUTO HELLO
+!: AUTO GOODBYE
+
+C: BEGIN {}
+S: SUCCESS {}
+C: RUN "RETURN 1" {} {}
+   PULL {"n": 1000}
+S: SUCCESS {}
+   FAILURE {"code": "#ERROR#", "message": "<whatever>"}
+C: RESET
+S: SUCCESS {}
+#EXTRA_RESET_1#
