+++ tests/stub/session_run_parameters/scripts/timeout.script
@@ -1,0 +1,10 @@
+!: BOLT 4
+!: AUTO HELLO
+!: AUTO GOODBYE
+!: AUTO RESET
+
+C: RUN "RETURN 1 as n" {} {"tx_timeout": 17}
+   PULL {"n": 1000}
+S: SUCCESS {"fields": ["n"]}
+   RECORD [1]
+   SUCCESS {"type": "w"}
