+++ tests/stub/session_run_parameters/scripts/tx_meta.script
@@ -1,0 +1,10 @@
+!: BOLT 4
+!: AUTO HELLO
+!: AUTO GOODBYE
+!: AUTO RESET
+
+C: RUN "RETURN 1 as n" {} {"tx_metadata": {"akey": "aval"}}
+   PULL {"n": 1000}
+S: SUCCESS {"fields": ["n"]}
+   RECORD [1]
+   SUCCESS {"type": "w"}
