+++ tests/stub/suites.py
@@ -7,9 +7,8 @@ import sys
 import unittest
 
 import tests.stub.retry as retry
-import tests.stub.sessionparameters as sessionparameters
-import tests.stub.transport as transport
-import tests.stub.txparameters as txparameters
+import tests.stub.transport.test_transport as transport
+import tests.stub.tx_begin_parameters.test_tx_begin_parameters as txparameters
 from tests.stub import authorization
 from tests.testenv import (
     begin_test_suite,
@@ -24,7 +23,6 @@ stub_suite = unittest.TestSuite()
 stub_suite.addTests(loader.loadTestsFromModule(authorization))
 stub_suite.addTests(loader.loadTestsFromModule(retry))
 stub_suite.addTests(loader.loadTestsFromModule(transport))
-stub_suite.addTests(loader.loadTestsFromModule(sessionparameters))
 stub_suite.addTests(loader.loadTestsFromModule(txparameters))
 
 stub_suite.addTest(loader.discover(
