+++ tests/stub/transport/scripts/reader_with_noops.script
@@ -1,0 +1,15 @@
+!: BOLT #BOLT_VERSION#
+!: AUTO RESET
+!: AUTO HELLO
+!: AUTO GOODBYE
+
+C: RUN "RETURN 1 as n" {} {}
+   PULL {"n": 1000}
+S: SUCCESS {"fields": ["n"]}
+   <NOOP>
+   <NOOP>
+   RECORD [1]
+   <NOOP>
+   <NOOP>
+   <NOOP>
+   SUCCESS {"type": "w"}
