+++ tests/stub/transport/test_transport.py
@@ -6,27 +6,8 @@ from tests.shared import (
 )
 from tests.stub.shared import StubServer
 
-script = """
-!: BOLT $bolt_version
-!: AUTO RESET
-!: AUTO HELLO
-!: AUTO GOODBYE
-
-C: RUN "RETURN 1 as n" {} {}
-   PULL { "n": 1000}
-S: SUCCESS {"fields": ["n"]}
-   <NOOP>
-   <NOOP>
-   RECORD [1]
-   <NOOP>
-   <NOOP>
-   <NOOP>
-   SUCCESS {"type": "w"}
-"""
-
-
 # Low-level network transport tests
-class Transport(TestkitTestCase):
+class TestTransport(TestkitTestCase):
     def setUp(self):
         super().setUp()
         self._server = StubServer(9001)
@@ -42,18 +23,21 @@ class Transport(TestkitTestCase):
         super().tearDown()
 
     def test_noop(self):
-        # Verifies that no op messages sent on bolt chunking layer are ignored. The no op messages
-        # are sent from server as a way to notify that the connection is still up.
+        # Verifies that no op messages sent on bolt chunking layer are ignored.
+        # The no op messages are sent from server as a way to notify that the
+        # connection is still up.
         # Bolt 4.1 >
         bolt_version = "4.1"
-        self._server.start(script=script, vars={"$bolt_version": bolt_version})
+        self._server.start(path=self.script_path("reader_with_noops.script"),
+                           vars={"#BOLT_VERSION#": bolt_version})
         result = self._session.run("RETURN 1 as n")
         record = result.next()
-        nilrec = result.next()
+        null_record = result.next()
         self._driver.close()
         self._server.done()
 
         # Verify the result
-        self.assertEqual(record.values[0].value, 1) # Indirectly verifies that we got a record
-        self.assertIsInstance(nilrec, types.NullRecord)
+        # Indirectly verifies that we got a record
+        self.assertEqual(record.values[0].value, 1)
+        self.assertIsInstance(null_record, types.NullRecord)
