+++ tests/stub/tx_begin_parameters/scripts/combined.script
@@ -1,0 +1,13 @@
+!: BOLT 4
+!: AUTO HELLO
+!: AUTO GOODBYE
+!: AUTO RESET
+
+C: BEGIN {"bookmarks": ["b0"], "tx_metadata": {"k": "v"}, "mode": "r", "tx_timeout": 11}
+S: SUCCESS {}
+C: RUN "RETURN 1 as n" {"p": 1} {}
+   PULL {"n": 1000}
+S: SUCCESS {"fields": ["n"]}
+   SUCCESS {"type": "r"}
+C: COMMIT
+S: SUCCESS {}
