+++ tests/stub/tx_begin_parameters/scripts/timeout.script
@@ -1,0 +1,13 @@
+!: BOLT 4
+!: AUTO HELLO
+!: AUTO GOODBYE
+!: AUTO RESET
+
+C: BEGIN {"tx_timeout": 17}
+S: SUCCESS {}
+C: RUN "RETURN 1 as n" {} {}
+   PULL {"n": 1000}
+S: SUCCESS {"fields": ["n"]}
+   SUCCESS {"type": "r"}
+C: COMMIT
+S: SUCCESS {}
